data/inputs/heritageai/requirements.txt
langchain==0.3.21
langchain-huggingface==0.1.2
langchain-openai==0.3.14
pymupdf==1.26.0
langchain_community==0.3.20
langgraph==0.3.11
tavily-python==0.7.3
streamlit==1.44.0
python-dotenv==1.1.0
plotly==6.0.1
langchain-chroma
openevals
    